+++ Source/Core/PerspectiveOffCenterFrustum.js
@@ -322,17 +322,19 @@ define([
      *
      * @param {Number} drawingBufferWidth The width of the drawing buffer.
      * @param {Number} drawingBufferHeight The height of the drawing buffer.
+     * @param {Number} pixelRatio The scaling factor from pixel space to coordinate space.
      * @param {Number} distance The distance to the near plane in meters.
      * @param {Cartesian2} result The object onto which to store the result.
      * @returns {Cartesian2} The modified result parameter or a new instance of {@link Cartesian2} with the pixel's width and height in the x and y properties, respectively.
      *
      * @exception {DeveloperError} drawingBufferWidth must be greater than zero.
      * @exception {DeveloperError} drawingBufferHeight must be greater than zero.
+     * @exception {DeveloperError} pixelRatio must be greater than zero.
      *
      * @example
      * // Example 1
      * // Get the width and height of a pixel.
-     * var pixelSize = camera.frustum.getPixelDimensions(scene.drawingBufferWidth, scene.drawingBufferHeight, 1.0, new Cesium.Cartesian2());
+     * var pixelSize = camera.frustum.getPixelDimensions(scene.drawingBufferWidth, scene.drawingBufferHeight, scene.pixelRatio, 1.0, new Cesium.Cartesian2());
      *
      * @example
      * // Example 2
@@ -343,9 +345,9 @@ define([
      * var toCenter = Cesium.Cartesian3.subtract(primitive.boundingVolume.center, position, new Cesium.Cartesian3());      // vector from camera to a primitive
      * var toCenterProj = Cesium.Cartesian3.multiplyByScalar(direction, Cesium.Cartesian3.dot(direction, toCenter), new Cesium.Cartesian3()); // project vector onto camera direction vector
      * var distance = Cesium.Cartesian3.magnitude(toCenterProj);
-     * var pixelSize = camera.frustum.getPixelDimensions(scene.drawingBufferWidth, scene.drawingBufferHeight, distance, new Cesium.Cartesian2());
+     * var pixelSize = camera.frustum.getPixelDimensions(scene.drawingBufferWidth, scene.drawingBufferHeight, scene.pixelRatio, distance, new Cesium.Cartesian2());
      */
-    PerspectiveOffCenterFrustum.prototype.getPixelDimensions = function(drawingBufferWidth, drawingBufferHeight, distance, result) {
+    PerspectiveOffCenterFrustum.prototype.getPixelDimensions = function(drawingBufferWidth, drawingBufferHeight, pixelRatio, distance, result) {
         update(this);
 
         //>>includeStart('debug', pragmas.debug);
@@ -358,6 +360,12 @@ define([
         if (drawingBufferHeight <= 0) {
             throw new DeveloperError('drawingBufferHeight must be greater than zero.');
         }
+        if (!defined(pixelRatio)) {
+            throw new DeveloperError('pixelRatio is required');
+        }
+        if (pixelRatio <= 0) {
+            throw new DeveloperError('pixelRatio must be greater than zero.');
+        }
         if (!defined(distance)) {
             throw new DeveloperError('distance is required.');
         }
@@ -368,9 +376,9 @@ define([
 
         var inverseNear = 1.0 / this.near;
         var tanTheta = this.top * inverseNear;
-        var pixelHeight = 2.0 * distance * tanTheta / drawingBufferHeight;
+        var pixelHeight = 2.0 * pixelRatio * distance * tanTheta / drawingBufferHeight;
         tanTheta = this.right * inverseNear;
-        var pixelWidth = 2.0 * distance * tanTheta / drawingBufferWidth;
+        var pixelWidth = 2.0 * pixelRatio * distance * tanTheta / drawingBufferWidth;
 
         result.x = pixelWidth;
         result.y = pixelHeight;