+++ Specs/Core/OrthographicOffCenterFrustumSpec.js
@@ -150,32 +150,50 @@ describe('Core/OrthographicOffCenterFrustum', function() {
 
     it('get pixel dimensions throws without canvas height', function() {
        expect(function() {
-            return frustum.getPixelDimensions(1.0, undefined, 0.0, new Cartesian2());
+            return frustum.getPixelDimensions(1.0, undefined, 1.0, 0.0, new Cartesian2());
        }).toThrowDeveloperError();
     });
 
     it('get pixel dimensions throws without canvas width', function() {
         expect(function() {
-            return frustum.getPixelDimensions(undefined, 1.0, 0.0, new Cartesian2());
+            return frustum.getPixelDimensions(undefined, 1.0, 1.0, 0.0, new Cartesian2());
         }).toThrowDeveloperError();
     });
 
     it('get pixel dimensions throws with canvas width less than or equal to zero', function() {
         expect(function() {
-            return frustum.getPixelDimensions(0.0, 1.0, 0.0, new Cartesian2());
+            return frustum.getPixelDimensions(0.0, 1.0, 1.0, 0.0, new Cartesian2());
         }).toThrowDeveloperError();
     });
 
     it('get pixel dimensions throws with canvas height less than or equal to zero', function() {
         expect(function() {
-            return frustum.getPixelDimensions(1.0, 0.0, 0.0, new Cartesian2());
+            return frustum.getPixelDimensions(1.0, 0.0, 1.0, 0.0, new Cartesian2());
+        }).toThrowDeveloperError();
+    });
+
+    it('get pixel dimensions throws without pixel ratio', function() {
+        expect(function() {
+            return frustum.getPixelDimensions(1.0, 1.0, undefined, 0.0, new Cartesian2());
+        }).toThrowDeveloperError();
+    });
+
+    it('get pixel dimensions throws with pixel ratio less than or equal to zero', function() {
+        expect(function() {
+            return frustum.getPixelDimensions(1.0, 1.0, 0.0, 0.0, new Cartesian2());
         }).toThrowDeveloperError();
     });
 
     it('get pixel dimensions', function() {
-        var pixelSize = frustum.getPixelDimensions(1.0, 1.0, 0.0, new Cartesian2());
+        var pixelSize = frustum.getPixelDimensions(1.0, 1.0, 1.0, 0.0, new Cartesian2());
         expect(pixelSize.x).toEqual(2.0);
         expect(pixelSize.y).toEqual(2.0);
+    });
+
+    it('get pixel dimensions with pixel ratio', function() {
+        var pixelSize = frustum.getPixelDimensions(1.0, 1.0, 2.0, 0.0, new Cartesian2());
+        expect(pixelSize.x).toEqual(4.0);
+        expect(pixelSize.y).toEqual(4.0);
     });
 
     it('equals', function() {